backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.2
requests==2.28.1
tabulate==0.8.10
werkzeug==2.0.3
numpy>=1.26.0
scikit-learn>=1.5.0